+++ publicodes/source/mecanisms/one-possibility.tsx
@@ -1,0 +1,9 @@
+// TODO : This isn't a real mecanism, cf. #963
+export const mecanismOnePossibility = dottedName => (recurse, k, v) => ({
+	...v,
+	'une possibilité': 'oui',
+	evaluate: (cache, situation, parsedRules, node) => ({
+		...node,
+		missingVariables: { [dottedName]: 1 }
+	})
+})
